Functioning with bounding box height and width
It looks like there is some improvement when objects are nearby/overlap
and have different sizes.  Overall, evaluation metrics look a bit worse
because objects can change size rapidly.  (This is kind of hacky, but
after the Kalman filter update I replace the state with measurement
values becaues of rapidly changing bounding boxes, without this
the results are much worse.)  It should be possible to improve
performance with bounding box size, but need to try something more,
a next step could be learning Q (process noise covariance) at different
regions through the image.  This could help if bounding boxes change
most rapidly when the cars parked along a street are leaving the camera's
field of view.