A note about the last commit
Actually the hack of replacing target states with measurement values
after the Kalman filter update produces results with quality almost
identical to without the hack (incorrectly stated the hack helped
a lot in the last message).  Visually looking at the results it appears
that both methods produce/eliminate different kinds of erros, so maybe
there is something to be gained here from a more principled approach.